server/requirements.txt
Authlib==1.3.1
Flask==3.0.3
PyJWT==2.8.0